Updated variable types

+++ Problems/74_Confusion_Matrix/problem.py
@@ -1,2 +1,2 @@
-def confusion_matrix(data: list[list[int|float]]) -> list[list[int|float]]:
+def confusion_matrix(data: list[list[int|int]]) -> list[list[int|int]]:
     return confusion_matrix
